When calling service-monitor.c:service_new, flush service_free_delayed.
service_free_delayed is a queue of recently exited, but not yet
cleaned up processes (i.e., not yet flushed from the service monitor's
data structures).  If a process exits and then is immediately
restarted, the old process may still be in the service_free_delayed
queue when service_new is called.  That means service-monitor thinks
the name is still taken.  To avoid this, call service_free_delayed_cb
in service_new to flush the service_free_delayed queue.